refactor(constitution): fix lint, types, docstrings, and test gaps

+++ packages/acgs-lite/src/acgs_lite/constitution/experience_library.py
@@ -430,4 +430,4 @@ def _cosine_sim(a: list[float], b: list[float]) -> float:
     mag_b = sum(x * x for x in b) ** 0.5
     if mag_a == 0.0 or mag_b == 0.0:
         return 0.0
-    return dot / (mag_a * mag_b)
+    return float(dot / (mag_a * mag_b))
